chore: WIP Make user.rs module
