+++ .idea/discord.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DiscordProjectSettings">
-    <option name="show" value="true" />
+    <option name="show" value="PROJECT_FILES" />
     <option name="nameOverrideEnabled" value="true" />
     <option name="nameOverrideText" value="Darakbang Engine" />
     <option name="description" value="The New Darakbang Engine powered by Rust" />
