+++ src/database/user/mod.rs
@@ -89,60 +89,4 @@ pub fn find_local_user(
     }
 }
 
-use serde::{Serialize, Deserialize};
-
-#[derive(Serialize, Deserialize)]
-pub struct TokenClaims {
-    exp: usize,
-    iat: usize,
-    iss: String,
-    sub: String,
-}
-
-use thiserror::Error;
-use crate::config;
-use crate::config::Config;
-
-#[derive(Error, Debug)]
-pub enum UserError {
-    #[error("User not found")]
-    NotFound,
-    #[error("Authentication error")]
-    Authentication,
-}
-
-pub fn auth_local_user(
-    conn: &MysqlConnection,
-    user_name: &str,
-    password: &str,
-) -> Result<String, UserError> {
-    let config: Config = config::Config::load().expect("Invalid configuration detected");
-    let user: User = match find_local_user(conn, user_name, password) {
-        Ok(Some(user)) => user,
-        _ => Err(UserError::NotFound)?,
-    };
-    let claim = TokenClaims {
-        exp: Utc::now().add(Duration::weeks(1)).timestamp() as usize,
-        iat: Utc::now().timestamp() as usize,
-        iss: config.jwt_issuer.clone(),
-        sub: user.uid,
-    };
-    encode(
-        &Header::default(),
-        &claim,
-        &EncodingKey::from_secret(config.jwt_secret.as_ref()),
-    ).map_err(|e| UserError::Authentication)
-}
-
-pub fn auth_token(
-    conn: &MysqlConnection,
-    token: &str,
-) -> Result<TokenClaims, UserError> {
-    let config: Config = config::Config::load().expect("Invalid configuration detected");
-    let token_data = decode::<TokenClaims>(
-        &token,
-        &DecodingKey::from_secret(config.jwt_secret.as_ref()),
-        &Validation::default(),
-    ).map_err(|e| UserError::Authentication)?;
-    Ok(token_data.claims)
-}
+pub mod token;
