+++ src/database/user/token/error/mod.rs
@@ -1,0 +1,2 @@
+mod user_auth;
+pub use user_auth::UserAuthError;
