+++ src/database/user/token/error/user_auth.rs
@@ -1,0 +1,9 @@
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum UserAuthError {
+    #[error("User not found")]
+    NotFound,
+    #[error("Authentication error")]
+    Authentication,
+}
